fix transaction after call sql

+++ Scripts/CreateTables.sql
@@ -3,12 +3,9 @@ USE [master];
 
 -- Drop database if it is already present
 IF EXISTS(
-  SELECT
-  *
-FROM
-  [sys].[databases]
-WHERE
-    [name] = 'telephone_company'
+  SELECT *
+  FROM [sys].[databases]
+  WHERE [name] = 'telephone_company'
 ) BEGIN
   DROP DATABASE [telephone_company]
 END
@@ -44,7 +41,7 @@ CREATE TABLE [accounts]
 (
   [account_id] INT IDENTITY(1, 1) PRIMARY KEY,
   [subscriber_id] INT NOT NULL,
-  [balance] INT DEFAULT 0,
+  [balance] DOUBLE PRECISION DEFAULT 0,
   FOREIGN KEY(subscriber_id) REFERENCES [subscribers](subscriber_id) ON DELETE CASCADE
 );
 
@@ -63,7 +60,7 @@ CREATE TABLE [transactions]
   [transaction_id] INT IDENTITY(1, 1) PRIMARY KEY,
   [transaction_type_id] INT NOT NULL,
   [account_id] INT NOT NULL,
-  [amount] INT NOT NULL,
+  [amount] DOUBLE PRECISION NOT NULL,
   [timestamp] DATETIME DEFAULT CURRENT_TIMESTAMP,
   FOREIGN KEY(transaction_type_id) REFERENCES [transaction_types](transaction_type_id) ON DELETE CASCADE,
   FOREIGN KEY(account_id) REFERENCES [accounts](account_id) ON DELETE CASCADE
@@ -101,7 +98,7 @@ CREATE TABLE [daytime_prices]
 (
   [price_id] INT NOT NULL,
   [daytime_id] INT NOT NULL,
-  [price_per_minute] INT NOT NULL,
+  [price_per_minute] FLOAT NOT NULL,
   PRIMARY KEY(price_id, daytime_id),
   FOREIGN KEY(price_id) REFERENCES [prices](price_id) ON DELETE CASCADE,
   FOREIGN KEY(daytime_id) REFERENCES [daytimes](daytime_id) ON DELETE CASCADE
